refactor: split rrule edge tests
